refactor: split bert into lib+bin, relocate ECS type aliases
Step 1 of #42: structural foundation for native library access.

- Add src/lib.rs that declares bevy_app and leptos_app modules with
  re-exports. main.rs becomes a thin shell calling bert::App.
- Move 9 Bevy ECS query/filter type aliases (SystemQuery,
  ExternalEntityFilter, etc.) from src/leptos_app/mod.rs into
  src/bevy_app/mod.rs. These are Bevy Query/Filter types composing
  ECS components — they had no Leptos-specific content and belonged
  on the Bevy side all along. They'd been placed on the Leptos side
  originally because that's where the first consumer was
  (single_query_signal<T, F> in leptos-bevy-canvas bridge).
- Disambiguate `components::System` from `bevy::prelude::System` trait
  in the SystemQuery and SubSystemQuery aliases.

This step only relocates; no feature flags yet. Both default cargo
check and cargo check --target wasm32-unknown-unknown pass. WASM +
Tauri + native library behaviors unchanged.

Next step: introduce `gui` feature flag to make WASM deps optional,
enabling native-only builds for the transpiler (#37).

Part of #42.
