+++ src/lib.rs
@@ -1,0 +1,9 @@
+#![feature(string_remove_matches)]
+#![allow(clippy::all)]
+#![allow(dead_code)]
+
+pub mod bevy_app;
+pub mod leptos_app;
+
+pub use bevy_app::*;
+pub use leptos_app::*;
