+++ src/main.rs
@@ -2,12 +2,8 @@
 #![allow(clippy::all)]
 #![allow(dead_code)]
 
-pub mod bevy_app;
-pub mod leptos_app;
-
-pub use bevy_app::*;
+use bert::App;
 use leptos::mount;
-pub use leptos_app::*;
 
 fn main() {
     console_error_panic_hook::set_once();
